pyproject.toml
[build-system]
requires = [
	"flit_core >=3.2,<4",
	"python_version >= '3.9'"
]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[project]
name = "husc"
authors = [{name = "Gyasu Bajracharya", email = "gyasu@outlook.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    "pytest",
	"numpy",
	"scikit-learn",
	"pandas",
	"matplotlib",
	"plotly",
	"biopython",  
	"tqdm",
	"statsmodels"        
]

[project.urls]
Home = "https://github.com/gyasu/HuSC.git"
